refactor: move model resolution to @mcp-cli/core, resolve in daemon too
Shortname resolution (opus→claude-opus-4-6 etc.) now lives in
packages/core/src/model.ts and is applied in both the CLI and daemon
paths, so MCP callers using claude_prompt directly also get resolution.
