+++ packages/command/src/commands/claude.ts
@@ -6,7 +6,7 @@
  */
 
 import { join } from "node:path";
-import { ipcCall } from "@mcp-cli/core";
+import { ipcCall, resolveModelName } from "@mcp-cli/core";
 import { c, printError as defaultPrintError, formatToolResult } from "../output";
 import { extractFullFlag, extractJsonFlag } from "../parse";
 
@@ -134,17 +134,8 @@ export async function cmdClaude(args: string[], deps?: Partial<ClaudeDeps>): Pro
 
 // ── Subcommands ──
 
-/** Map short model names to full model IDs. */
-export const MODEL_SHORTNAMES: Record<string, string> = {
-  opus: "claude-opus-4-6",
-  sonnet: "claude-sonnet-4-6",
-  haiku: "claude-haiku-4-5-20251001",
-};
-
-/** Resolve a model name: accept shortnames or pass through full IDs. */
-export function resolveModelName(input: string): string {
-  return MODEL_SHORTNAMES[input.toLowerCase()] ?? input;
-}
+// Re-export for tests
+export { MODEL_SHORTNAMES, resolveModelName } from "@mcp-cli/core";
 
 export interface SpawnArgs {
   task: string | undefined;
