+++ packages/core/src/index.ts
@@ -7,3 +7,4 @@ export * from "./constants";
 export * from "./env";
 export * from "./fs";
 export * from "./schema-display";
+export * from "./model";
